// src/interfaces/address/created.address.dto.ts
export interface ICreatedAddressDto {
    ownerAddressId: number;
    idPerson: number;
    idAdress: number;
    cnpj: string;
    isDefault: boolean;
    ownerName: string;
    ownerPhone: string;
    ownerPhone2: string;
    ownerEmail: string;
    updateDate: string;
    id: number;
    active: boolean;
    createDate: string;
    cep: string;
    city: string;
    street: string;
    number: string;
    complement: string;
    state: string;
    district: string;
}